+++ include/odf/res/pal.h
@@ -11,7 +11,7 @@ OptionalPtr* palOpen(InMemoryFile* file);
 
 ucvec3 palGetColor(Palette* palette, uint8_t index);
 
-void palUnindex(Palette* palette, ucvec4* buffer, bool isTransparent, uint8_t* indexedStream, uint32_t length);
+void palUnindex(Palette* palette, ucvec4* buffer, bool isTransparent, uint8_t* indexedStream, size_t length);
 
 void palClose(Palette* pal);
